Composite similarity between vectors

Compute composite measures, GCSM or CMSC, between two vectors.

cmsc(
  x,
  y,
  rescale = FALSE,
  xmin = NA_real_,
  xmax = NA_real_,
  ymin = NA_real_,
  ymax = NA_real_,
  comp = "si"
)

cmsc_e1(
  x,
  y,
  rescale = FALSE,
  xmin = NA_real_,
  xmax = NA_real_,
  ymin = NA_real_,
  ymax = NA_real_,
  comp = "si"
)

cmsc_e2(
  x,
  y,
  rescale = FALSE,
  xmin = NA_real_,
  xmax = NA_real_,
  ymin = NA_real_,
  ymax = NA_real_,
  comp = "si"
)

gcsm(
  x,
  y,
  rescale = FALSE,
  xmin = NA_real_,
  xmax = NA_real_,
  ymin = NA_real_,
  ymax = NA_real_,
  comp = "si"
)

Arguments

x
A vector.

y
The other vector.

rescale
Rescale or not before computation.

xmin, xmax, ymin, ymax
Normalization parameters. If NA, are calculated from the ranges of x and y, respectively. See Details.

comp
Variable to return. If "si", the composite measure, if "s1","s2" or "s3", the corresponding component.

Value

A number.

Details

These functions compute composite measures between vectors. Missing values are omitted. Normalization parameters are used to rescale x and y, and determine the global minimum (min) and maximum (max). If rescale is TRUE, x and y are rescaled to (x-xmin)/(xmax-xmin) and (y-ymin)/(ymax-ymin); and set min=0, max=1. If FALSE, min=min(xmin,ymin), max=max(xmax,ymax).

Examples

x = runif(9)
gcsm(x, x)

#> [1] 1

cmsc(x, x)

#> [1] 1

# mean shift
gcsm(x, x - 0.2, xmin = 0, xmax = 1, ymin = 0, ymax = 1)

#> [1] 0.8

cmsc(x, x - 0.2, xmin = 0, xmax = 1, ymin = 0, ymax = 1)

#> [1] 0.96

gcsm(x, x + 0.2, xmin = 0, xmax = 1, ymin = 0, ymax = 1)

#> [1] 0.8

cmsc(x, x + 0.2, xmin = 0, xmax = 1, ymin = 0, ymax = 1)

#> [1] 0.96

## dissimilarity
y = 1 - x # y is the perfect antianalog of x
gcsm(y, x)

#> [1] -1

gcsm(y, x - 0.2, xmin = 0, xmax = 1, ymin = 0, ymax = 1)

#> [1] -0.8

gcsm(y, x + 0.2, xmin = 0, xmax = 1, ymin = 0, ymax = 1)

#> [1] -0.8

# random noise
noise = rnorm(9, mean = 0, sd = 0.2)
gcsm(x, x + noise, xmin = 0, xmax = 1, ymin = 0, ymax = 1)

#> [1] 0.6430002

cmsc(x, x + noise, xmin = 0, xmax = 1, ymin = 0, ymax = 1)

#> [1] 0.8005446

## dissimilarity
gcsm(y, x + noise, xmin = 0, xmax = 1, ymin = 0, ymax = 1)

#> [1] -0.6430002