Composite similarity on spatial windows

Compute composite measures, GCSM, CMSC or SSIM, on spatial windows.

cmsc_sw(
  x,
  y,
  rescale = FALSE,
  xmin = NA_real_,
  xmax = NA_real_,
  ymin = NA_real_,
  ymax = NA_real_,
  ksize = 9,
  globe = FALSE,
  comp = "si"
)

cmsc_e1_sw(
  x,
  y,
  rescale = FALSE,
  xmin = NA_real_,
  xmax = NA_real_,
  ymin = NA_real_,
  ymax = NA_real_,
  ksize = 9,
  globe = FALSE,
  comp = "si"
)

cmsc_e2_sw(
  x,
  y,
  rescale = FALSE,
  xmin = NA_real_,
  xmax = NA_real_,
  ymin = NA_real_,
  ymax = NA_real_,
  ksize = 9,
  globe = FALSE,
  comp = "si"
)

gcsm_sw(
  x,
  y,
  rescale = FALSE,
  xmin = NA_real_,
  xmax = NA_real_,
  ymin = NA_real_,
  ymax = NA_real_,
  ksize = 9,
  globe = FALSE,
  comp = "si"
)

ssim_sw(
  x,
  y,
  rescale = FALSE,
  xmin = NA_real_,
  xmax = NA_real_,
  ymin = NA_real_,
  ymax = NA_real_,
  ksize = 11,
  sigma = 1.5,
  globe = FALSE,
  comp = "si"
)

Arguments

x
A matrix.

y
The other matrix.

rescale
Rescale or not before computation.

xmin, xmax, ymin, ymax
Normalization parameters. If NA, are calculated from the ranges of x and y, respectively. See Details.

ksize
Side length of spatial windows.

globe
Are data at the global scale? If TRUE, two vertical borders will be padded before computation.

comp
Variable to return. If "si", the composite measure, if "s1","s2" or "s3", the corresponding component.

sigma
Standard deviation of Gaussian weighting function depending on the distance between the cell and kernel center.

Value

A matrix.

Details

These functions slide the spatial window over space. Missing values are omitted. Normalization parameters are used to rescale x and y, and determine the global minimum (min) and maximum (max). If rescale is TRUE, x and y are rescaled to (x-xmin)/(xmax-xmin) and (y-ymin)/(ymax-ymin); and set min=0, max=1. If FALSE, min=min(xmin,ymin), max=max(xmax,ymax). OpenMP is used for parallel computing.

Examples

x = matrix(runif(36), nrow = 6, ncol = 6)

gcsm_sw(x, x + 0.2, xmin = 0, xmax = 1, ymin = 0, ymax = 1, ksize = 3)

#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]  0.8  0.8  0.8  0.8  0.8  0.8
#> [2,]  0.8  0.8  0.8  0.8  0.8  0.8
#> [3,]  0.8  0.8  0.8  0.8  0.8  0.8
#> [4,]  0.8  0.8  0.8  0.8  0.8  0.8
#> [5,]  0.8  0.8  0.8  0.8  0.8  0.8
#> [6,]  0.8  0.8  0.8  0.8  0.8  0.8

cmsc_sw(x, x + 0.2, xmin = 0, xmax = 1, ymin = 0, ymax = 1, ksize = 3)

#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,] 0.96 0.96 0.96 0.96 0.96 0.96
#> [2,] 0.96 0.96 0.96 0.96 0.96 0.96
#> [3,] 0.96 0.96 0.96 0.96 0.96 0.96
#> [4,] 0.96 0.96 0.96 0.96 0.96 0.96
#> [5,] 0.96 0.96 0.96 0.96 0.96 0.96
#> [6,] 0.96 0.96 0.96 0.96 0.96 0.96

ssim_sw(x, x + 0.2, xmin = 0, xmax = 1, ymin = 0, ymax = 1, ksize = 3)

#>           [,1]      [,2]      [,3]      [,4]      [,5]      [,6]
#> [1,] 0.9585344 0.9514243 0.9568176 0.9464671 0.9390995 0.9053013
#> [2,] 0.9582709 0.9528179 0.9600679 0.9565130 0.9445835 0.9027066
#> [3,] 0.9472250 0.9521555 0.9673027 0.9662072 0.9559058 0.9260077
#> [4,] 0.9408251 0.9528877 0.9686123 0.9672505 0.9632375 0.9533851
#> [5,] 0.9195519 0.9362902 0.9554898 0.9513427 0.9529162 0.9501868
#> [6,] 0.9170119 0.9199597 0.9302770 0.9268874 0.9345928 0.9355408